Refactor project layout and document architecture
