+++ src/main.rs
@@ -1,6 +1,6 @@
-use vpscheck::http;      // ← 库 crate 的名字就是 package name
+use vpscheck::web;      // ← 库 crate 的名字就是 package name
 
 #[tokio::main]
 async fn main() -> std::io::Result<()> {
-    http::run().await
+    web::run().await
 }
